backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
psycopg[binary]==3.2.13
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
